requirements.txt
Django==3.2.9
djangorestframework==3.13.1
Pillow==8.4.0
drf-spectacular==0.21.0